Refresh documentation and move Ray tooling to contrib
