+++ docs/source/migration/migrating_from_omniisaacgymenvs.rst
@@ -12,10 +12,15 @@ We have updated OmniIsaacGymEnvs to Isaac Sim version 4.0.0 to support the migra
 to Isaac Lab. Moving forward, OmniIsaacGymEnvs will be deprecated and future development
 will continue in Isaac Lab.
 
-.. note::
+.. important::
 
-  The following changes are with respect to Isaac Lab 1.0 release. Please refer to the `release notes`_ for any changes
-  in the future releases.
+   The detailed source-framework comparison below was originally written for Isaac Lab 1.0.
+   Use it to translate OmniIsaacGymEnvs concepts, then apply the
+   :ref:`migrating-to-isaaclab-3-0` guide for the current multi-backend configuration,
+   ``ProxyArray`` access, quaternion ordering, preset selection, and indexed or masked write APIs.
+   The maintained Cartpole implementation in
+   :isaaclab-source:`isaaclab_tasks/core/cartpole <source/isaaclab_tasks/isaaclab_tasks/core/cartpole>`
+   is the canonical current example.
 
 Task Config Setup
 ~~~~~~~~~~~~~~~~~
@@ -67,7 +72,7 @@ be specified on a per-actor basis in the config class for each individual articu
 When running simulation on the GPU, buffers in PhysX require pre-allocation for computing and storing
 information such as contacts, collisions and aggregate pairs. These buffers may need to be adjusted
 depending on the complexity of the environment, the number of expected contacts and collisions,
-and the number of actors in the environment. The :class:`~isaaclab.sim.PhysxCfg` class provides access
+and the number of actors in the environment. The :class:`~isaaclab_physx.physics.PhysxCfg` class provides access
 for setting the GPU buffer dimensions.
 
 +--------------------------------------------------+---------------------------------------------------------------+
@@ -88,7 +93,7 @@ for setting the GPU buffer dimensions.
 || dynamic_friction: 1.0                           || dynamic_friction=1.0,                                        |
 || restitution: 0.0                                || restitution=0.0                                              |
 ||                                                 || )                                                            |
-|| physx:                                          || physx: PhysxCfg = PhysxCfg(                                  |
+|| physx:                                          || physics=PhysxCfg(                                             |
 || worker_thread_count: ${....num_threads}         || # worker_thread_count is no longer needed                    |
 || solver_type: ${....solver_type}                 || solver_type=1,                                               |
 || use_gpu: ${contains:"cuda",${....sim_device}}   || # use_gpu is deduced from the device                         |
