+++ docs/source/refs/bibliography.rst
@@ -1,4 +1,14 @@
 Bibliography
 ============
 
+Isaac Lab
+---------
+
 .. bibliography::
+   :filter: key == "mittal2025isaaclab"
+
+Related Work
+------------
+
+.. bibliography::
+   :filter: key != "mittal2025isaaclab"
