+++ source/isaaclab/changelog.d/doc-cleanup-0816.rst
@@ -1,0 +1,3 @@
+Added
+^^^^^
+* Added the ``ray`` optional dependency group for the contributed Ray job-dispatch and tuning workflow.
