+++ source/isaaclab_contrib/changelog.d/doc-cleanup-0816.rst
@@ -1,0 +1,6 @@
+Changed
+^^^^^^^
+
+* Moved the contributed Ray job-dispatch and tuning workflow from
+  ``scripts/reinforcement_learning/ray`` to ``isaaclab_contrib.rl.ray``. Run its tools with
+  ``uv run --extra ray python -m isaaclab_contrib.rl.ray.<tool>``.
