+++ source/isaaclab_contrib/isaaclab_contrib/rl/ray/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""Contributed Ray job-dispatch and hyperparameter-tuning workflows."""
