+++ source/isaaclab_contrib/isaaclab_contrib/rl/ray/hyperparameter_tuning/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""Example Ray Tune configurations."""
