+++ source/isaaclab_contrib/isaaclab_contrib/rl/ray/launch.py
@@ -9,18 +9,18 @@ Usage:
 
 .. code-block:: bash
     # If the head node is stuck on container creating, make sure to create a secret
-    python3 scripts/reinforcement_learning/ray/launch.py -h
+    uv run --extra ray python -m isaaclab_contrib.rl.ray.launch -h
 
     # Examples
 
     # The following creates 8 GPUx1 nvidia l4 workers
-    python3 scripts/reinforcement_learning/ray/launch.py --cluster_host google_cloud \
+    uv run --extra ray python -m isaaclab_contrib.rl.ray.launch --cluster_host google_cloud \
     --namespace <NAMESPACE> --image <YOUR_ISAAC_RAY_IMAGE> \
     --num_workers 8 --num_clusters 1 --worker_accelerator nvidia-l4 --gpu_per_worker 1
 
     # The following creates 1 GPUx1 nvidia l4 worker, 2 GPUx2 nvidia-tesla-t4 workers,
     # and 2 GPUx4 nvidia-tesla-t4 GPU workers
-    python3 scripts/reinforcement_learning/ray/launch.py --cluster_host google_cloud \
+    uv run --extra ray python -m isaaclab_contrib.rl.ray.launch --cluster_host google_cloud \
     --namespace <NAMESPACE> --image <YOUR_ISAAC_RAY_IMAGE> \
     --num_workers 1 2 --num_clusters 1 \
     --worker_accelerator nvidia-l4 nvidia-tesla-t4 --gpu_per_worker 1 2 4
@@ -34,8 +34,7 @@ import yaml
 from jinja2 import Environment, FileSystemLoader
 from kubernetes import config
 
-# Local imports
-import util  # isort: skip
+from isaaclab_contrib.rl.ray import util
 
 RAY_DIR = pathlib.Path(__file__).parent
