+++ source/isaaclab_contrib/isaaclab_contrib/rl/ray/task_runner.py
@@ -52,10 +52,10 @@ Typical usage:
 .. code-block:: bash
 
     # Print help and argument details:
-    python task_runner.py -h
+    uv run --extra ray python -m isaaclab_contrib.rl.ray.task_runner -h
 
     # Submit tasks defined in a YAML file to the Ray cluster (auto-detects Ray head address):
-    python task_runner.py --task_cfg /path/to/tasks.yaml
+    uv run --extra ray python -m isaaclab_contrib.rl.ray.task_runner --task_cfg /path/to/tasks.yaml
 
 YAML configuration example-1:
 -----------------------------
@@ -113,8 +113,7 @@ from datetime import datetime
 
 import yaml
 
-# Local imports
-import util  # isort: skip
+from isaaclab_contrib.rl.ray import util
 
 # Safe operators for arithmetic expression evaluation
 _SAFE_OPERATORS = {
